firehose_service/requirements.txt
asyncio==4.0.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
db-dtypes==1.4.4
debugpy
decorator
dotenv==0.9.9
executing
gevent==25.9.1
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
matplotlib-inline
nest_asyncio
numpy==2.3.4
packaging
pandas==2.3.3
parso
pexpect
platformdirs
prompt_toolkit
proto-plus==1.26.1
protobuf==6.33.0
psutil
ptyprocess
pure_eval
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments
python-dateutil
python-dotenv==1.2.1
pytz==2025.2
pyzmq
requests==2.32.5
rsa==4.9.1
six
stack_data
tornado
traitlets
typing_extensions
tzdata==2025.2
urllib3==2.5.0
wcwidth
websockets==15.0.1
zipp
zope.event==6.0
zope.interface==8.0.1